+++ src/main/java/be/kdg/int5/GameSDK.java
@@ -1,11 +1,14 @@
 package be.kdg.int5;
 
+import be.kdg.int5.domain.Achievement;
 import be.kdg.int5.domain.GameContext;
+import be.kdg.int5.domain.Rule;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.math.BigDecimal;
 import java.net.http.HttpClient;
 import java.time.Instant;
+import java.util.List;
 import java.util.Objects;
 
 public class GameSDK {
@@ -48,7 +51,10 @@ public class GameSDK {
             String description,
             BigDecimal price,
             String iconUrl,
-            String backgroundUrl
+            String backgroundUrl,
+            List<Rule> rules,
+            List<String> screenshots,
+            List<Achievement> achievements
     ) {
         return RegisterGameModule.registerGame(
                 this,
@@ -57,7 +63,10 @@ public class GameSDK {
                 description,
                 price,
                 iconUrl,
-                backgroundUrl
+                backgroundUrl,
+                rules,
+                screenshots,
+                achievements
         );
     }