+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Each change should be sorted in to one of the following categories:
 
 ## [Unreleased]
 ### Added
-- registerGame sdk method (without any collections)
+- registerGame sdk method
 - authentication using api key to the GameSDK
 - GameSDK.Builder to build a new GameSDK object
 ### Changed